Test : 통합 테스트 코드 작성

+++ src/main/java/cjkimhello97/toy/crashMyServer/click/domain/Click.java
@@ -31,7 +31,7 @@ public class Click implements Comparable<Click> {
     private Long clickId;
 
     @Column(name = "count")
-    private Double count;
+    private Integer count;
 
     @OneToOne(fetch = LAZY)
     @JoinColumn(name = "member_id")
